Upgrade properties from `attributes.py` to library for things we recommend as current best-practice
