+++ src/atopile/attributes.py
@@ -331,16 +331,6 @@ class Resistor(F.Resistor):
         _handle_package_shim(self, value, "R")
 
     @property
-    def p1(self) -> F.Electrical:
-        """One side of the resistor."""
-        return self.unnamed[0]
-
-    @property
-    def p2(self) -> F.Electrical:
-        """The other side of the resistor."""
-        return self.unnamed[1]
-
-    @property
     def _1(self) -> F.Electrical:
         return self.unnamed[0]
 
@@ -361,19 +351,6 @@ class CommonCapacitor(F.Capacitor):
     """
     These attributes are common to both electrolytic and non-electrolytic capacitors.
     """
-
-    class _has_power(L.Trait.decless()):
-        """
-        This trait is used to add power interfaces to
-        capacitors who use them, keeping the interfaces
-        off caps which don't use it.
-
-        Caps have power-interfaces when used with them.
-        """
-
-        def __init__(self, power: F.ElectricPower) -> None:
-            super().__init__()
-            self.power = power
 
     @property
     def value(self):
@@ -419,16 +396,6 @@ class CommonCapacitor(F.Capacitor):
         GlobalAttributes.footprint.fset(self, value)
 
     @property
-    def p1(self) -> F.Electrical:
-        """One side of the capacitor."""
-        return self.unnamed[0]
-
-    @property
-    def p2(self) -> F.Electrical:
-        """The other side of the capacitor."""
-        return self.unnamed[1]
-
-    @property
     def _1(self) -> F.Electrical:
         return self.unnamed[0]
 
@@ -452,18 +419,6 @@ class Capacitor(CommonCapacitor):
         trait.pinmap["2"] = self.p2
         return trait
 
-    @property
-    def power(self) -> F.ElectricPower:
-        """A `Power` interface, which is connected to the capacitor."""
-        if self.has_trait(self._has_power):
-            power = self.get_trait(self._has_power).power
-        else:
-            power = F.ElectricPower()
-            power.hv.connect_via(self, power.lv)
-            self.add(self._has_power(power))
-
-        return power
-
 
 @_register_shim(
     "generics/capacitors.ato:CapacitorElectrolytic", "import CapacitorElectrolytic"
@@ -476,6 +431,7 @@ class CapacitorElectrolytic(CommonCapacitor):
 
     pickable = None
 
+    # Overrides the default implementation in F.Capacitor
     @property
     def power(self) -> F.ElectricPower:
         if self.has_trait(self._has_power):
@@ -495,16 +451,6 @@ class Inductor(F.Inductor):
     This inductor is replaces `generics/inductors.ato:Inductor`
     every times it's referenced.
     """
-
-    @property
-    def p1(self) -> F.Electrical:
-        """Signal to one side of the inductor."""
-        return self.unnamed[0]
-
-    @property
-    def p2(self) -> F.Electrical:
-        """Signal to the other side of the inductor."""
-        return self.unnamed[1]
 
     @property
     def _1(self) -> F.Electrical:
@@ -548,16 +494,6 @@ class LED(F.LED):
 @_register_shim("generics/interfaces.ato:Power", "import ElectricPower")
 class Power(F.ElectricPower):
     """Temporary shim to translate `value` to `power`."""
-
-    @property
-    def vcc(self) -> F.Electrical:
-        """Higher-voltage side of the power interface."""
-        return self.hv
-
-    @property
-    def gnd(self) -> F.Electrical:
-        """Lower-voltage side of the power interface."""
-        return self.lv
 
     @property
     def current(self):
